monkey patch run_commands and use it it in conda ops create

+++ src/commands.py
@@ -4,14 +4,30 @@ from pathlib import Path
 import sys
 import configparser
 import json
+from ruamel.yaml import YAML
+from .split_requirements import create_split_files
+from .python_api import run_command
+import conda.cli.python_api
+
+import logging
+
+logger = logging.getLogger("conda.cli.python_api")
+logger.setLevel(logging.DEBUG)
+
+ch = logging.StreamHandler()
+ch.setFormatter(logging.Formatter("%(asctime)s %(name)-15s %(levelname)-8s %(processName)-10s %(message)s"))
+logger.addHandler(ch)
 
 CONDA_OPS_DIR_NAME = '.conda-ops'
 STATUS_FILENAME = 'status.txt'
 CONFIG_FILENAME = 'config.ini'
-REQUIREMENTS_FILENAME = 'requirements'
+REQUIREMENTS_FILENAME = 'environment.yml'
 LOCK_FILENAME = 'lockfile'
 
-
+yaml = YAML()
+yaml.default_flow_style=False
+yaml.width=4096
+yaml.indent(offset=4)
 
 def ops_init():
     '''
@@ -33,6 +49,8 @@ def ops_init():
     # setup initial config
     config_file = conda_ops_path / CONFIG_FILENAME
     config = configparser.ConfigParser()
+
+    # currently defaults to creating an env_name based on the location of the project
     env_name = Path.cwd().name
     config['DEFAULT'] = {'ENV_NAME': env_name}
     with open(config_file, 'w') as f:
@@ -40,9 +58,16 @@ def ops_init():
 
     # create basic requirements file
     requirements_file = conda_ops_path / REQUIREMENTS_FILENAME
-    requirements_dict = {'env_name': env_name, 'dependencies': ['python', 'pip']}
-    with open(requirements_file, 'w') as f:
-        json.dump(requirements_dict, f)
+    if requirements_file.exists():
+        requirements_dict = {'name': env_name,
+                             'channels': ['defaults'],
+                             'channel-order': ['defaults'],
+                             'dependencies': ['python', 'pip']}
+        print('rewriting')
+        with open(requirements_file, 'w') as f:
+            yaml.dump(requirements_dict, f)
+    else:
+        print(f'Requirements file {requirements_file} already exists')
     print(f'Initialized conda-ops project in {conda_ops_path.resolve()}')
 
 def ops_create():
@@ -51,16 +76,41 @@ def ops_create():
     '''
     print('TODO: check if the environment already exists...')
 
+    print(conda.cli.python_api.__name__)
+    print(logger)
     ops_dir = find_conda_ops_dir()
+    requirements_file = ops_dir / REQUIREMENTS_FILENAME
+
+    requirements = yaml.load(requirements_file)
+    env_name = requirements['name']
+    print(f'creating the environment {env_name}')
+
+    print('generating multi-step requirements files')
+    create_split_files(requirements_file, ops_dir)
+
+    with open(ops_dir / '.ops.channel-order.include', 'r') as f:
+        order_list = f.read().split()
+
+    with open(ops_dir / f'.ops.{order_list[0]}-environment.txt') as f:
+        package_list = f.read().split()
+
+    create_args = ["-n", env_name] + package_list + ['--dry-run', '--json']
+
+    print(create_args)
+
+    stdout, stderr, result_code = run_command("create", create_args, use_exception_handler=True)
+    if result_code != 0:
+        print(stderr)
+        sys.exit()
+    json_reqs = json.loads(stdout)
+    print(json_reqs.keys())
+    print(json_reqs)
 
     print('generating the lock file')
     lock_file = ops_dir / LOCK_FILENAME # we may not want this until we create the first environment
     lock_file.touch()
 
-    config = configparser.ConfigParser()
-    config.read(ops_dir / CONFIG_FILENAME)
-    env_name = config['DEFAULT']['ENV_NAME']
-    print(f'creating the environment {env_name}')
+    print('TODO: continuing to the other steps...')
 
     status_file = ops_dir / STATUS_FILENAME
     with open(status_file, 'w') as f:
